Re-wrote proxy.php to use 'wget' external command rather than 'readfile()' on a URL.
The 'readfile()' command, when used over HTTP on a URL,
seems to send a very basic HTTP request.  The web server
or something on the new web host, Network Solutions, is
rejecting this request, returning a 400 HTTP response
("Bad Request").  Tried using PHP sockets, couldn't get
that to work. So had to rely on the in-elegant (but at
least it's functional) wget method.